tazawodAcademy/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
kombu==5.3.7
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.6
six==1.16.0
SQLAlchemy==2.0.31
sqlparams==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
